Add memoryfilesystem as a test dependency
FIXME: not sure why, but none of the dependencies in the current pom.xml are
marked as <scope>test</scope>, so this added dependency does not specify it
either. My maven-fu is not good enough to tell whether this is a problem at all.

Comment as to why this dependency is needed; we cannot mock a Path's existence,
unlike what is possible for File (which is not even final!).